do not exec the out binary in out script
[fixes #97783956]

This occasionally caused the `ssh-agent` process to become orphaned as
it's parent would become the Go process which exits but still has an
unexpected child process.

The strange thing was that this was non-deterministic and only caused
Garden to block on `Wait()` quite rarely (although when it did it was a
pain to recover from). There's a theory that this is a Garden bug due to
a process not being considered completed until its standard streams are
closed. It's possible that the `ssh-agent` was holding these open.