Triggering suppression
Sometimes it is not desirable to trigger a job when modifying a pool resource.
This commit adds an optional parameter to suppress triggering. The default is
that triggering may occur, which is compatible with the behaviour before this
change was introduced.

Note: the generate statement for fake_lock_handler.go is updated to use the
correct output directory.