Add atomic update operation
If the existing lock is in the unclaimed state we will update it with
the contents of the `metadata` file. If no such lock is present in
either the claimed or unclaimed state we add a new lock to the pool in
the unclaimed state. If the lock is in the claimed state we will wait
for it to be unclaimed and proceed to update it as above.

- Update existing fakes using counterfeiter
- Add fake for UpdateLock
- Add unit and integration tests
- Update README

Signed-off-by: Kaiting Chen <kaitingc@vmware.com>